lib: Move stats implementation to oci.Runtime
The way stats are retrieved is tied to the way the runtime is being
handled, in this case assuming the containers are running on the host.

Because we want to better support VM based container runtimes, it is
necessary to make sure we don't make such assumption from the lib
package since it is high level.

Fixes #1946
