+++ lib/container_server_unsupported.go
@@ -15,8 +15,3 @@ func (c *ContainerServer) addSandboxPlatform(sb *sandbox.Sandbox) {
 func (c *ContainerServer) removeSandboxPlatform(sb *sandbox.Sandbox) {
 	// nothin' doin'
 }
-
-func (c *ContainerServer) getContainerStats(ctr *oci.Container, previousStats *ContainerStats) (*ContainerStats, error) {
-	// nothin' doin'
-	return nil, errors.New("container stats not supported")
-}
